new structure of layout and classes, step 1
