+++ refactored_modular/interscalehub_manager.py
